V2 Rewrite (#28)
This is essentially a rewrite of the library, using https://github.com/jbreckmckye/electron-login-test to test that everything still works.

The new version uses a library-adapter architecture I've been exploring in a few projects. The idea is to make it easy to augment the library with user-defined plugins, or even alternative auth providers (making this a general Electron PKCE library).